Refactor: RP-28 시큐리티 파일 구조 변경
