+++ src/main/java/com/cherrypick/backend/domain/oauth2/dto/OAuth2LoginSuccessResponseDTO.java
@@ -1,0 +1,14 @@
+package com.cherrypick.backend.domain.oauth2.dto;
+
+import lombok.Builder;
+
+@Builder
+public record OAuth2LoginSuccessResponseDTO(
+        Long userId,
+        String accessToken,
+        String refreshToken,
+        String redirectURL,
+        boolean isNewUser
+
+) {
+}
