+++ src/main/java/com/cherrypick/backend/global/config/SecurityConfig.java
@@ -1,9 +1,10 @@
 package com.cherrypick.backend.global.config;
 
-import com.cherrypick.backend.global.config.oauth.*;
+import com.cherrypick.backend.global.config.security.CustomAuthorizationRequestResolver;
+import com.cherrypick.backend.domain.oauth2.service.OAuth2Service;
+import com.cherrypick.backend.global.config.security.OAuth2SuccessHandler;
+import com.cherrypick.backend.global.config.security.*;
 import com.cherrypick.backend.global.util.JWTUtil;
-import jakarta.servlet.http.HttpServletRequest;
-import kotlin.io.AccessDeniedException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -13,22 +14,13 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.oauth2.client.registration.ClientRegistrationRepository;
-import org.springframework.security.oauth2.client.web.DefaultOAuth2AuthorizationRequestResolver;
-import org.springframework.security.oauth2.client.web.OAuth2AuthorizationRequestResolver;
-import org.springframework.security.oauth2.core.endpoint.OAuth2AuthorizationRequest;
-import org.springframework.security.web.FilterChainProxy;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 @Configuration
 @EnableWebSecurity @RequiredArgsConstructor
